Fix eslint issues
