+++ src/components/screens/txDetail/lskSummary.js
@@ -21,6 +21,72 @@ const txTypes = [
   'vote',
 ];
 
+const getConfig = (styles, tx, accountAddress) => {
+  if (accountAddress !== tx.senderAddress && tx.type === 0) {
+    return {
+      arrowStyle: styles.reverseArrow,
+      amountStyle: [styles.incoming, styles.theme.incoming],
+      firstAddress: tx.recipientAddress,
+      secondAddress: tx.senderAddress,
+      amountSign: '',
+      direction: 'incoming',
+    };
+  }
+  return {
+    arrowStyle: null,
+    amountStyle: [styles.outgoing, styles.theme.outgoing],
+    firstAddress: tx.senderAddress,
+    secondAddress: tx.recipientAddress,
+    amountSign: '-',
+    direction: 'outgoing',
+  };
+};
+
+const Graphics = ({
+  styles,
+  tx,
+  theme,
+  config,
+}) => (
+  <View style={styles.row}>
+    {tx.type !== 0 || tx.recipientAddress === tx.senderAddress ? (
+      <Image
+        style={{ width: 40, height: 40 }}
+        source={transactions[txTypes[tx.type]].image(theme)}
+      />
+    ) : (
+      <Fragment>
+        <Avatar address={config.firstAddress} size={40} />
+        {theme === themes.light ? (
+          <Image source={arrowLight} style={[styles.arrow, config.arrowStyle]} />
+        ) : (
+          <Image source={arrowDark} style={[styles.arrow, config.arrowStyle]} />
+        )}
+        <Avatar address={config.secondAddress} size={40} />
+      </Fragment>
+    )}
+  </View>
+);
+
+const TimeStamp = ({
+  timestamp,
+  styles,
+}) => {
+  if (timestamp) {
+    return (
+      <FormattedDate
+          format="MMM D, YYYY LTS"
+          type={P}
+          style={[styles.date, styles.theme.date]}
+        >
+        {timestamp}
+      </FormattedDate>
+    );
+  }
+
+  return null;
+};
+
 const LskSummary = ({
   styles,
   theme,
@@ -31,50 +97,24 @@ const LskSummary = ({
   language,
 }) => {
   const amount = fromRawLsk(tx.amount);
-  let arrowStyle;
-  let amountStyle = [styles.outgoing, styles.theme.outgoing];
-  let firstAddress = tx.senderAddress;
-  let secondAddress = tx.recipientAddress;
-  let amountSign = '-';
-  let direction = 'outgoing';
-
-  if (accountAddress !== tx.senderAddress && tx.type === 0) {
-    arrowStyle = styles.reverseArrow;
-    amountStyle = [styles.incoming, styles.theme.incoming];
-    firstAddress = tx.recipientAddress;
-    secondAddress = tx.senderAddress;
-    amountSign = '';
-    direction = 'incoming';
-  }
+  const config = getConfig(styles, tx, accountAddress);
 
   return (
     <View style={[styles.senderAndRecipient, styles.theme.senderAndRecipient]}>
-      <View style={styles.row}>
-        {tx.type !== 0 || tx.recipientAddress === tx.senderAddress ? (
-          <Image
-            style={{ width: 40, height: 40 }}
-            source={transactions[txTypes[tx.type]].image(theme)}
-          />
-        ) : (
-          <Fragment>
-            <Avatar address={firstAddress} size={40} />
-            {theme === themes.light ? (
-              <Image source={arrowLight} style={[styles.arrow, arrowStyle]} />
-            ) : (
-              <Image source={arrowDark} style={[styles.arrow, arrowStyle]} />
-            )}
-            <Avatar address={secondAddress} size={40} />
-          </Fragment>
-        )}
-      </View>
+      <Graphics
+        styles={styles}
+        tx={tx}
+        theme={theme}
+        config={config}
+      />
       {tx.type !== 0 || tx.recipientAddress === tx.senderAddress ? (
-        <H3 style={amountStyle}>{t(transactions[txTypes[tx.type]].title)}</H3>
+        <H3 style={config.amountStyle}>{t(transactions[txTypes[tx.type]].title)}</H3>
       ) : null}
       {tx.type === 0
       && tx.recipientAddress !== tx.senderAddress
       && !incognito ? (
-        <H3 style={amountStyle}>
-          {amountSign}
+        <H3 style={config.amountStyle}>
+          {config.amountSign}
           <FormattedNumber language={language}>
             {fromRawLsk(tx.amount)}
           </FormattedNumber>
@@ -83,17 +123,12 @@ const LskSummary = ({
       {tx.type === 0
       && tx.recipientAddress !== tx.senderAddress
       && incognito ? (
-        <Blur value={amount} direction={direction} style={styles.amountBlur} />
+        <Blur value={amount} direction={config.direction} style={styles.amountBlur} />
         ) : null}
-      {tx.timestamp ? (
-        <FormattedDate
-          format="MMM D, YYYY LTS"
-          type={P}
-          style={[styles.date, styles.theme.date]}
-        >
-          {tx.timestamp}
-        </FormattedDate>
-      ) : null}
+      <TimeStamp
+        timestamp={tx.timestamp}
+        styles={styles}
+      />
     </View>
   );
 };
